refactor(cli,daemon): put the policy with the consumer, the empty list with the daemon
Two corrections to the restructure, both about which layer owns a decision.

`announce.rs` had grown a `use crate::shard_cmd::{version_meets_minimum,
BlockServerEntry}` — the wire format importing from the code that reads it,
which is the inversion the restructure exists to remove. The split is by kind,
not by proximity: `ServerInfoFields` and `decode_server_info_ext` are bytes to
fields, including the `addrs_signed` verification, and stay beside
`to_msgpack`; `decode_server_info_regular` and `_dictionary` decide that
`state != 2` is not worth chaining and that a pre-0.3.15 record is not worth
trusting, which is sharding's opinion and nobody else's, so they go back to
`shard_cmd.rs` on top of `decode_server_info_ext`. `version_meets_minimum` is
private again. The tests follow their functions: the round trips, the forged
signer, the unparseable envelope and the legacy cases test the wire format and
stay; the two dictionary subkey-binding tests test the policy and move, taking
`dictionary_bytes` and `signed_envelope` with them.

The other finding was the same three-line fallback standing at three call
sites, because `do_connect` refused a CONNECT with no addresses as malformed.
It is not malformed: `host.Connect(AddrInfo{ID, Addrs: []})` on a Go routed
host consults routing, and that is exactly the case a caller has when the peer
record it read published no addresses. `do_connect` now takes a valid peer with
an empty list down the bare-`/p2p/<id>` path into `dispatch_routed`, so its
three shapes are routed, classic and seeded. Each call site is one
`connect_peer_with_addrs` again, and the failure a caller sees for a peer
nothing knows about is "peer not found in DHT" rather than a complaint about
its own request.
